// web/src/shared/api/generated/model/getQuestionsParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { QuestionType } from './questionType';
import type { QuestionValidationStatus } from './questionValidationStatus';

export type GetQuestionsParams = {
status?: QuestionValidationStatus;
type?: QuestionType;
subjectId?: string;
lessonId?: string;
teacherId?: string;
/**
 * @pattern ^-?(?:0|[1-9]\d*)$
 */
minVersion?: number | string;
rejectionReason?: string;
/**
 * @pattern ^-?(?:0|[1-9]\d*)$
 */
pageNumber?: number | string;
/**
 * @pattern ^-?(?:0|[1-9]\d*)$
 */
pageSize?: number | string;
};